Create You have been insulted
This script is a .py script that is supposed to take random Monty Python insults and propagate them in an alert box any time you close a window. I was bored and was sick of MacAfee popping up on start up so I wanted to create a script where when I closed MacAfee it insulted me with a Monty Python quote. Maybe someone might find this useful or entertaining. I had issues getting it to work with system.